requirements.txt
numpy
gymnasium
matplotlib
torch
ipython